fixing models and endpoint auth

+++ backend/src/app.py
@@ -1,9 +1,13 @@
 import os
 from flask import Flask, request, jsonify, abort
 from flask_cors import CORS
+from dotenv import load_dotenv
+import os
 
-from .database.models import db_drop_and_create_all, setup_db, RSVP, Invitation
-from .auth.auth import AuthError, requires_auth
+load_dotenv()
+
+from database.models import db_drop_and_create_all, setup_db, RSVP, Invitation
+from auth.auth import AuthError, requires_auth
 
 def create_app(test_config=None):
 
@@ -12,9 +16,13 @@ def create_app(test_config=None):
     CORS(app)
     db_drop_and_create_all()
 
+    @app.route('/invitations')
+    def get_invitations():
+        invitations = Invitation.query.all()
+        return jsonify([i.format() for i in invitations])
+
     @app.route('/invitations/<int:id>')
-    @requires_auth('get:invitation')
-    def get_invitation(payload, id):
+    def get_invitation(id):
         invitation = Invitation.query.filter(Invitation.id == id).one_or_none()
         if invitation:
             return jsonify(invitation.format())
@@ -25,13 +33,19 @@ def create_app(test_config=None):
     @requires_auth('get:rsvps')
     def get_rsvps(payload, invitation_id):
         rsvps = RSVP.query.filter_by(invitation_id=invitation_id).all()
-        return jsonify([r.format() for r in rsvps])
+        return jsonify(sucess=True, rsvp=[r.format() for r in rsvps])
 
     @app.route('/invitations/<int:invitation_id>/rsvps/<int:rsvp_id>')
     @requires_auth('get:rsvp')
     def get_rsvp(payload, invitation_id, rsvp_id):
         rsvp = RSVP.query.filter(RSVP.invitation_id==invitation_id, RSVP.id==rsvp_id).one_or_none()
         if rsvp:
+            if rsvp.jwt_sub != payload['sub']:
+                raise AuthError({
+                    'code': 'unauthorized',
+                    'description': 'permissions are not included in the payload'
+                    }, 401
+                )
             return jsonify(rsvp.format())
         else:
             abort(404)
@@ -43,8 +57,8 @@ def create_app(test_config=None):
         try:
             name = data['name']
             email = data['email']
-            plus_one = data['plus_one']
-            invitation = Invitation(name=name, email=email, plus_one=plus_one)
+            text = data['text']
+            invitation = Invitation(name=name, email=email, text=text)
             invitation.insert()
             return jsonify(invitation.format())
         except:
@@ -62,8 +76,8 @@ def create_app(test_config=None):
                 invitation.name = data['name']
             if 'email' in data:
                 invitation.email = data['email']
-            if 'plus_one' in data:
-                invitation.plus_one = data['plus_one']
+            if 'text' in data:
+                invitation.text = data['text']
             invitation.update()
             return jsonify(invitation.format())
         except:
@@ -83,7 +97,8 @@ def create_app(test_config=None):
             abort(500)
     
     @app.route('/invitations/<int:invitation_id>/rsvps', methods=['POST'])
-    def create_rsvp(invitation_id):
+    @requires_auth('post:rsvp')
+    def create_rsvp(payload, invitation_id):
         invitation = Invitation.query.get(invitation_id)
         if invitation is None:
             abort(404)
@@ -97,7 +112,7 @@ def create_app(test_config=None):
             if not guest_name or not guest_email or not response:
                 abort(404)
 
-            rsvp = RSVP(guest_name=guest_name, guest_email=guest_email, response=response, invitation=invitation)
+            rsvp = RSVP(guest_name=guest_name,jwt_sub=payload['sub'], guest_email=guest_email, response=response, invitation_id=invitation_id)
             rsvp.insert()
 
             return jsonify(rsvp.format())
@@ -106,16 +121,23 @@ def create_app(test_config=None):
 
 
     @app.route('/invitations/<int:invitation_id>/rsvps/<int:rsvp_id>', methods=['DELETE'])
-    def delete_rsvp(invitation_id, rsvp_id):
+    @requires_auth('delete:rsvp')
+    def delete_rsvp(payload, invitation_id, rsvp_id):
         invitation = Invitation.query.get(invitation_id)
         if invitation is None:
             abort(404)
 
         try:
-
             rsvp = RSVP.query.filter_by(invitation=invitation, id=rsvp_id).first()
             if rsvp is None:
                 abort(404)
+
+            if rsvp.jwt_sub != payload['sub']:
+                raise AuthError({
+                    'code': 'unauthorized',
+                    'description': 'permissions are not included in the payload'
+                    }, 401
+                )
 
             rsvp.delete()
 
@@ -125,7 +147,8 @@ def create_app(test_config=None):
 
 
     @app.route('/invitations/<int:invitation_id>/rsvps/<int:rsvp_id>', methods=['PATCH'])
-    def update_rsvp(invitation_id, rsvp_id):
+    @requires_auth('patch:rsvp')
+    def update_rsvp(payload, invitation_id, rsvp_id):
         invitation = Invitation.query.get(invitation_id)
         if invitation is None:
             abort(404)
@@ -134,6 +157,13 @@ def create_app(test_config=None):
             rsvp = RSVP.query.filter_by(invitation=invitation, id=rsvp_id).first()
             if rsvp is None:
                 abort(404)
+
+            if rsvp.jwt_sub != payload['sub']:
+                raise AuthError({
+                    'code': 'unauthorized',
+                    'description': 'permissions are not included in the payload'
+                    }, 401
+                )
 
             data = request.get_json()
             guest_name = data.get('guest_name')
@@ -199,7 +229,7 @@ def create_app(test_config=None):
             'message': error.error['description']
         }), error.status_code
 
-        return app
+    return app
 
 app = create_app()
